Update README.md
Merged build and usage as both seem to be the same thing just two ways to get there as most Github users have Git.

Note: Linux GTK Client was listed as a Key Socket server. Neither really mean anything for users and are implementation details. Renamed to daemon to mean that it's just a program running in background (no matter client or server).