Add 22tracks, Birp, Bugs, Spreaker, and Streamsquid support
These additions were PRs that were rejected since the websites weren't popular enough, however I feel that this extension should provide as much compatibility as possible.  As much as I hate to say it, it's not popular enough for many sites to incorporate the event API method just yet.  I've tested these and confirmed that they function as expected.  :)