Support Deezer's new interface
This query supports the new Deezer interface in beta (which is miles better), and is backward-compatible with the current interface at the same time.

The change is they removed the `.control-*` class from all control buttons for some reason, so all of them have the same set of classes now. That's why I opted for using `aria-label` in queries.

The class `.player-controls` at the beginning can be ommitted (tested in both interfaces, each query still returns the same single element), but I thought it's safer to keep it.